srcCodeBD/myFirstClass/requirements.txt
flask
flask_sqlalchemy
flask_swagger_ui
pyjwt
